ai_recruitment_backend/requirements.txt
langgraph
langchain
langchain-community
dashscope
pydantic
python-dotenv

